test(e2e) add test for upgrades
Add a new test to deploy the previous version of KIC manifests and then
the current version of KIC manifests to confirm that upgrades are
viable.

Split verifyIngress() into deployIngress() and verifyIngress().
deployIngress() creates the test Ingress and associated resources.
verifyIngress() confirms the resulting route is available on the proxy.

Add a KongIngress to the deployIngress() and verifyIngress() functions
to perform a basic test of our CRDs.

Add helper functions to retrieve the current and previous Git tag per
the local repo.